requirements.txt
amqp==5.1.1
annotated-types==0.6.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.78
botocore==1.29.78
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==1.2.0
Django==4.2.5
django-admin-env-notice==1.0
django-admin-interface==0.24.2
django-celery-results==2.4.0
django-colorfield==0.8.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-heroku==0.3.1
django-ninja==1.1.0
django-storages==1.13.2
dnspython==2.6.1
email_validator==2.1.1
Faker==12.0.1
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
mccabe==0.7.0
mixer==7.2.2
Pillow==9.4.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.0.1
Pygments==2.14.0
pyhumps==3.8.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.10.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
